+++ modules/Team/teamController.ts
@@ -4,5 +4,5 @@ import { createCrudController } from "../../helper/crud";
 export const teamController = createCrudController(TeamModel, {
   label: "Team",
   populate: [{ path: "league", model: "League" }],
-  fields: ["name", "logo", "league"],
+  fields: ["name", "logo", "league", "short_name"],
 });